fix: review

+++ src/features/delete-notice/api/fetch-delete-notice.ts
@@ -2,7 +2,7 @@ import { apiClient } from "@/shared/libs/custom-axios";
 
 export const fetchDeleteNotice = async (id: string) => {
   try {
-    const { data } = await apiClient.delete(`/api/notices/${id}`);
+    const { data } = await apiClient.delete(`/api/admin/notices/${id}`);
     return data;
   } catch (error) {
     console.error("notice delete api", error);
